COM-1301 rename stack to AppContainer + other fix
