+++ app.json
@@ -20,7 +20,7 @@
     "assetBundlePatterns": ["assets/*"],
     "ios": {
       "bundleIdentifier": "com.alenvi.compani",
-      "buildNumber": "1.0.3"
+      "buildNumber": "1.0.2"
     },
     "android": {
       "package": "com.alenvi.compani",
